+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 4.6.1 - ??
+## [4.7.0] - 2016-09-10
+ - New Topics provider
+ - added getRelatedTopics to Interests provider
  - Provider::execGetRequest method now returns bool or array
  - Exceptions classes renamed
  - Helpers classes renamed